+++ backend/.gitignore
@@ -26,6 +26,7 @@ node_modules/
 ## Fordítási fájlok
 bin/
 obj/
+.vs/
 
 ## Unit Test fájlok
 TestResults/
